Delete detector concept; trainable models are the only model surface
Implements docs/plans/delete-detectors.md. Single concept on disk
(`data/trainable_models/<name>.json`), single in-memory cache
(`DetectorContext`), and single resolve→embed→train pipeline
(`train_from_labelset`). MLP weights live only in RAM.

Backend
- Delete `vtsearch/models/weights_compat.py`,
  `vtsearch/processors/`, `vtsearch/routes/processor_importers.py`,
  `vtsearch/routes/detectors_*.py`.
- Replace with `vtsearch/routes/processors.py` (extractors / localizers
  / pregen processors only), `vtsearch/routes/model_scoring.py`
  (auto-detect + find-label backed by trainable-model labelsets), and
  `vtsearch/routes/model_find.py` (multi-dataset Find).
- Drop the `autorun_detectors` global, the `add/remove/rename/get_*_autorun_detector`
  helpers, and the `weights` field on every JSON shape.
- Drop the `autorun_processors`, `autorun_detector_names`, and
  `detectors_dir` settings; everything routes through
  `autorun_trainable_models` now.
- Model registry: drop `trainable: bool`, `detector_name`, and
  `trainable_model_name` fields. Every entry is a trainable model;
  `name` is the labelset filename.
- Add `PUT /api/models/registry/<id>/autorun` (writes
  `autorun_trainable_models`), include an `autorun` flag in the
  registry list response.
- CLI `--autodetect` only loads trainable models from settings; no
  more detector JSON files.

Frontend
- Drop `DetectorInfo`, `Autorun*`, `Detector*Response`,
  `DetectorServerFilesResponse`, `AutorunProcessor` from
  api.models.ts. Add `autorun` flag to `ModelRegistryEntry`.
- Trim `DetectorsApiService` down to autorun toggle,
  extractors / localizers, /api/auto-detect, /api/find-label,
  /api/find{,/check-labels,/progress}, and pregen processors.
- LoadSortModal: drop the "Server Detectors" section, emit a
  `modelSelected` event with the registry id; LabelView consumes
  it via `/api/find-label`.
- DetectorStateService no longer tracks detectors.
- Dashboard `toggleAutorun` calls the new registry endpoint.

Cleanup
- Delete `tests/test_processor_importers.py`; remove
  `vtsearch/processors/importers/server_detector_file/requirements.txt`
  from `requirements-plugins.txt`.

Tests are next; this commit gets the build green so changes are
recoverable.
