+++ frontend/src/app/components/label-view/label-view.component.html
@@ -26,7 +26,7 @@
       (textSort)="onTextSort($event)"
       (learnedSort)="onLearnedSort(false)"
       (loadSort)="onLoadSort()"
-      (detectorLoaded)="onDetectorLoaded($event)"
+      (modelSelected)="onModelSelected($event)"
       (exampleSortStarted)="onExampleSortStarted($event)"
       (mediaSelect)="onMediaSelect($event)"
       (mediaVote)="onHoverVote($event)"
