+++ frontend/src/app/components/left-panel/left-panel.component.html
@@ -87,7 +87,7 @@
           (textSort)="textSort.emit($event)"
           (learnedSort)="learnedSort.emit()"
           (loadSort)="loadSort.emit()"
-          (detectorLoaded)="detectorLoaded.emit($event)"
+          (modelSelected)="modelSelected.emit($event)"
           (exampleSortStarted)="exampleSortStarted.emit($event)"
         />
