+++ frontend/src/app/services/detector-state.service.ts
@@ -1,17 +1,19 @@
 import { Injectable, OnDestroy } from '@angular/core';
 import { BehaviorSubject, Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
-import { DetectorInfo } from '../models/api.models';
 import { DetectorsApiService } from './detectors-api.service';
 
+/**
+ * Shared cache for autorun extractors / localizers loaded from the backend.
+ * Detectors no longer exist as a separate concept — model state lives in
+ * the model registry instead (see TrainableModelsApiService).
+ */
 @Injectable({ providedIn: 'root' })
 export class DetectorStateService implements OnDestroy {
-  private readonly detectorsSubject = new BehaviorSubject<DetectorInfo[]>([]);
   private readonly extractorsSubject = new BehaviorSubject<unknown[]>([]);
   private readonly localizersSubject = new BehaviorSubject<unknown[]>([]);
   private readonly destroy$ = new Subject<void>();
 
-  readonly detectors$ = this.detectorsSubject.asObservable();
   readonly extractors$ = this.extractorsSubject.asObservable();
   readonly localizers$ = this.localizersSubject.asObservable();
 
@@ -22,23 +24,12 @@ export class DetectorStateService implements OnDestroy {
     this.destroy$.complete();
   }
 
-  get detectors(): DetectorInfo[] {
-    return this.detectorsSubject.value;
-  }
-
   get extractors(): unknown[] {
     return this.extractorsSubject.value;
   }
 
   get localizers(): unknown[] {
     return this.localizersSubject.value;
-  }
-
-  loadDetectors(): void {
-    this.detectorsApi
-      .getAutorunDetectors()
-      .pipe(takeUntil(this.destroy$))
-      .subscribe((resp) => this.detectorsSubject.next(resp.detectors));
   }
 
   loadExtractors(): void {
@@ -56,13 +47,11 @@ export class DetectorStateService implements OnDestroy {
   }
 
   loadAll(): void {
-    this.loadDetectors();
     this.loadExtractors();
     this.loadLocalizers();
   }
 
   clear(): void {
-    this.detectorsSubject.next([]);
     this.extractorsSubject.next([]);
     this.localizersSubject.next([]);
   }
