--- vtsearch/processors/__init__.py
@@ -1,1 +1,0 @@
-"""Processor modules: importers for creating autorun processors (detectors, extractors)."""
