--- vtsearch/processors/importers/server_detector_file/requirements.txt
@@ -1,1 +1,0 @@
-# Server detector file importer — no additional dependencies beyond core.
