+++ vtsearch/routes/__init__.py
@@ -8,12 +8,13 @@ from vtsearch.routes.media_server import media_server_bp
 from vtsearch.routes.medias import medias_bp
 from vtsearch.routes.datasets import datasets_bp
 from vtsearch.routes.datasets_registry import datasets_registry_bp
-from vtsearch.routes.detectors import detectors_bp
 from vtsearch.routes.exporters import exporters_bp
 from vtsearch.routes.label_importers import label_importers_bp
 from vtsearch.routes.main import main_bp
+from vtsearch.routes.model_find import model_find_bp
+from vtsearch.routes.model_scoring import model_scoring_bp
 from vtsearch.routes.models_registry import models_registry_bp
-from vtsearch.routes.processor_importers import processor_importers_bp
+from vtsearch.routes.processors import processors_bp
 from vtsearch.routes.settings import settings_bp
 from vtsearch.routes.settings_io import settings_io_bp
 from vtsearch.routes.sorting import sorting_bp
@@ -30,13 +31,14 @@ __all__ = [
     "medias_bp",
     "sorting_bp",
     "sync_sources_bp",
-    "detectors_bp",
     "datasets_bp",
     "datasets_registry_bp",
     "exporters_bp",
     "label_importers_bp",
+    "model_find_bp",
+    "model_scoring_bp",
     "models_registry_bp",
-    "processor_importers_bp",
+    "processors_bp",
     "settings_bp",
     "settings_io_bp",
     "trainable_models_bp",
