--- vtsearch/routes/detectors.py
@@ -1,32 +1,0 @@
-"""Blueprint for detector, extractor, and localizer routes.
-
-This module was split into sub-blueprints.  It re-exports them so that
-existing code that imports ``detectors_bp`` continues to work.
-"""
-
-from __future__ import annotations
-
-from flask import Blueprint
-
-from vtsearch.routes.detectors_crud import (  # noqa: F401
-    SERVER_DETECTOR_DIR,
-    get_detectors_dir,
-    _EXTRACTOR_FACTORIES,
-    _LOCALIZER_FACTORIES,
-    _build_extractor,
-    _build_localizer,
-    _ensure_extractor_factories,
-    _ensure_localizer_factories,
-    detectors_crud_bp,
-)
-from vtsearch.routes.detectors_find import detectors_find_bp  # noqa: F401
-from vtsearch.routes.detectors_scoring import detectors_scoring_bp  # noqa: F401
-from vtsearch.routes.detectors_training import detectors_training_bp  # noqa: F401
-
-detectors_bp = Blueprint("detectors", __name__)
-
-# Register sub-blueprints so all routes appear under the single detectors_bp.
-detectors_bp.register_blueprint(detectors_crud_bp)
-detectors_bp.register_blueprint(detectors_training_bp)
-detectors_bp.register_blueprint(detectors_find_bp)
-detectors_bp.register_blueprint(detectors_scoring_bp)
