+++ vtsearch/utils/__init__.py
@@ -33,7 +33,6 @@ from vtsearch.utils.state import (
     unregister_detector_context,
     with_dataset_context,
     with_detector_context,
-    add_autorun_detector,
     add_autorun_extractor,
     add_autorun_localizer,
     add_label_to_history,
@@ -56,7 +55,6 @@ from vtsearch.utils.state import (
     diversity_tree_label,
     diversity_tree_next_sample,
     diversity_tree_unlabel,
-    autorun_detectors,
     autorun_extractors,
     autorun_localizers,
     find_missing_entries,
@@ -64,9 +62,6 @@ from vtsearch.utils.state import (
     get_dupe_count,
     get_calibration_fraction,
     get_diversity_tree,
-    get_autodetect_detectors_by_media,
-    get_autorun_detectors,
-    get_autorun_detectors_by_media,
     get_autorun_extractors,
     get_autorun_extractors_by_media,
     get_autorun_localizers,
@@ -83,13 +78,8 @@ from vtsearch.utils.state import (
     last_learned_scores,
     next_media_id,
     remove_click_time,
-    remove_autorun_detector,
     remove_autorun_extractor,
-    set_autorun_detector_autodetect,
-    set_autorun_detector_examples,
-    get_autorun_detector_examples,
     remove_autorun_localizer,
-    rename_autorun_detector,
     rename_autorun_extractor,
     rename_autorun_localizer,
     resolve_media_ids,
@@ -122,7 +112,6 @@ __all__ = [
     "good_votes",
     "bad_votes",
     "label_history",
-    "autorun_detectors",
     "autorun_extractors",
     "autorun_localizers",
     "vote_click_times",
@@ -155,15 +144,6 @@ __all__ = [
     "textsort_suggestions",
     "add_textsort_suggestion",
     "get_textsort_suggestions",
-    "add_autorun_detector",
-    "remove_autorun_detector",
-    "rename_autorun_detector",
-    "get_autodetect_detectors_by_media",
-    "get_autorun_detectors",
-    "get_autorun_detectors_by_media",
-    "set_autorun_detector_autodetect",
-    "set_autorun_detector_examples",
-    "get_autorun_detector_examples",
     "add_autorun_extractor",
     "remove_autorun_extractor",
     "rename_autorun_extractor",
