+++ vtsearch/utils/state_processors.py
@@ -1,4 +1,4 @@
-"""Autorun processor (detector, extractor, localizer) CRUD operations."""
+"""Autorun extractor / localizer CRUD operations."""
 
 from __future__ import annotations
 
@@ -6,171 +6,9 @@ from typing import Any
 
 from vtsearch.utils.state_core import (
     _state_lock,
-    autorun_detectors,
     autorun_extractors,
     autorun_localizers,
 )
-
-
-# ---------------------------------------------------------------------------
-# Detectors
-# ---------------------------------------------------------------------------
-
-
-def add_autorun_detector(
-    name: str,
-    media_type: str,
-    weights: dict[str, Any] | None = None,
-    threshold: float = 0.5,
-    *,
-    autodetect: bool = False,
-    examples: list[dict[str, str]] | None = None,
-    num_labels: int = 0,
-    created_by: str = "default",
-    good_origins: list[dict[str, Any]] | None = None,
-    bad_origins: list[dict[str, Any]] | None = None,
-    inclusion: int = 0,
-) -> None:
-    """Add or overwrite a named autorun detector in the global store.
-
-    If a detector with the same ``name`` already exists it is replaced.
-
-    Args:
-        name: Unique human-readable name for the detector (e.g. ``"dog barks"``).
-        media_type: The media type the detector was trained on (``"audio"``,
-            ``"video"``, ``"image"``, or ``"text"``).
-        weights: Dict mapping layer-parameter names (e.g. ``"0.weight"``) to
-            lists of float values, representing the serialised MLP state dict.
-            May be ``None`` for an untrained detector stub.
-        threshold: Decision boundary score in ``[0, 1]``. Clips scoring at or
-            above this value are classified as positive.  Defaults to ``0.5``.
-        autodetect: Whether this detector is included when running
-            autodetect.  Defaults to ``False``.
-        examples: Optional list of example dicts, each with ``"type"``
-            (``"text"``, ``"media"``, or ``"detector"``) and ``"value"`` (str).
-        num_labels: Number of training labels used when this detector was last
-            trained.  Defaults to ``0`` for untrained stubs.
-        created_by: Username of the user who created this detector.
-        good_origins: List of origin dicts for media voted Good, each with
-            ``"origin"``, ``"origin_name"``, ``"filename"``, ``"md5"`` keys.
-            Used to re-derive weights without serialising them to disk.
-        bad_origins: List of origin dicts for media voted Bad (same shape).
-        inclusion: The inclusion value used when training this detector.
-    """
-    import time
-
-    with _state_lock:
-        autorun_detectors[name] = {
-            "name": name,
-            "media_type": media_type,
-            "weights": weights,
-            "threshold": threshold,
-            "created_at": time.time(),
-            "autodetect": autodetect,
-            "examples": examples or [],
-            "num_labels": num_labels,
-            "created_by": created_by,
-            "good_origins": good_origins or [],
-            "bad_origins": bad_origins or [],
-            "inclusion": inclusion,
-        }
-
-
-def remove_autorun_detector(name: str) -> bool:
-    """Remove a named autorun detector from the global store.
-
-    Returns:
-        ``True`` if the detector was found and removed; ``False`` if no
-        detector with that name exists.
-    """
-    with _state_lock:
-        if name in autorun_detectors:
-            del autorun_detectors[name]
-            return True
-        return False
-
-
-def rename_autorun_detector(old_name: str, new_name: str) -> bool:
-    """Rename a autorun detector, updating its internal ``"name"`` field.
-
-    The operation is atomic with respect to the dict: the old entry is removed
-    and a new entry is created in a single step (no window where neither exists).
-
-    Returns:
-        ``True`` if the rename succeeded (old name existed and new name was not
-        already taken); ``False`` otherwise (no changes are made).
-    """
-    with _state_lock:
-        if old_name in autorun_detectors and new_name not in autorun_detectors:
-            autorun_detectors[new_name] = autorun_detectors[old_name].copy()
-            autorun_detectors[new_name]["name"] = new_name
-            del autorun_detectors[old_name]
-            return True
-        return False
-
-
-def get_autorun_detectors() -> dict[str, dict[str, Any]]:
-    """Return a shallow copy of all autorun detectors."""
-    with _state_lock:
-        return autorun_detectors.copy()
-
-
-def get_autorun_detectors_by_media(media_type: str) -> dict[str, dict[str, Any]]:
-    """Return all autorun detectors matching a given media type."""
-    with _state_lock:
-        return {name: det for name, det in autorun_detectors.items() if det["media_type"] == media_type}
-
-
-def set_autorun_detector_autodetect(name: str, autodetect: bool) -> bool:
-    """Set the autodetect flag on a named autorun detector.
-
-    Returns:
-        ``True`` if the detector was found and updated; ``False`` otherwise.
-    """
-    with _state_lock:
-        if name in autorun_detectors:
-            autorun_detectors[name]["autodetect"] = autodetect
-            return True
-        return False
-
-
-def set_autorun_detector_examples(name: str, examples: list[dict[str, str]]) -> bool:
-    """Set the examples list on a named autorun detector.
-
-    Returns:
-        ``True`` if the detector was found and updated; ``False`` otherwise.
-    """
-    with _state_lock:
-        if name in autorun_detectors:
-            autorun_detectors[name]["examples"] = examples
-            return True
-        return False
-
-
-def get_autorun_detector_examples(name: str) -> list[dict[str, str]]:
-    """Return the examples list for a named autorun detector.
-
-    Returns an empty list if the detector is not found or has no examples.
-    """
-    with _state_lock:
-        det = autorun_detectors.get(name)
-        if det is None:
-            return []
-        return list(det.get("examples") or [])
-
-
-def get_autodetect_detectors_by_media(media_type: str) -> dict[str, dict[str, Any]]:
-    """Return autorun detectors matching a media type with autodetect enabled.
-
-    Like :func:`get_autorun_detectors_by_media` but also filters to only
-    include detectors whose ``"autodetect"`` flag is ``True``.
-    """
-    with _state_lock:
-        return {
-            name: det
-            for name, det in autorun_detectors.items()
-            if det["media_type"] == media_type and det.get("autodetect", True) and det.get("weights")
-        }
 
 
 # ---------------------------------------------------------------------------
